atividades de C
treinando alguma das atividades mais fáceis, pois são as que mais eu quebro cabeça kkkkk, e estruturas de repetição que está sendo meio dificil de fixar, mas vou.

+++ 1CursoC/atividades/atividade92.c
@@ -4,7 +4,7 @@
 int main(){
 
     int res;
-    unsigned long long soma = 1;
+    int soma = 1;
 
     printf("Digite um valor: ");
     scanf("%d", &res);
@@ -15,7 +15,7 @@ int main(){
     }
     
     
-    printf("%llu", soma);
+    printf("%d", soma);
 
     return 0;
 }
